Move types to Pkg.ts
