+++ scripts/commands/updateApiDocs.ts
@@ -32,7 +32,8 @@ import { specialCaseResults } from "../lib/api/specialCaseResults";
 import addFrontMatter from "../lib/api/addFrontMatter";
 import { dedupeHtmlIdsFromResults } from "../lib/api/dedupeHtmlIds";
 import { removePrefix, removeSuffix } from "../lib/stringUtils";
-import { Pkg, PkgInfo, Link, getPkgRoot } from "../lib/sharedTypes";
+import { Link } from "../lib/sharedTypes";
+import { Pkg, PkgInfo, getPkgRoot } from "../lib/api/Pkg";
 import { zxMain } from "../lib/zx";
 import { pathExists, getRoot } from "../lib/fs";
 import { downloadCIArtifact } from "../lib/api/downloadCIArtifacts";
