+++ scripts/lib/api/releaseNotes.ts
@@ -13,7 +13,7 @@
 import { $ } from "zx";
 import { getRoot, pathExists } from "../fs";
 import { parse } from "path";
-import { Pkg } from "../sharedTypes";
+import { Pkg } from "./Pkg";
 import { readFile, writeFile, readdir } from "fs/promises";
 
 interface releaseNoteEntry {
